requirements_dev.txt
pylint
pylint-django
flake8
ipython
pytest
pytest-django